requirements.txt
beautifulsoup4
flask
flask_wtf
wtforms
pymongo